+++ src/main/kotlin/at/orchaldir/gm/utils/renderer/RenderOptions.kt
@@ -3,7 +3,7 @@ package at.orchaldir.gm.utils.renderer
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class LineOptions(val color: RenderColor, val width: UInt)
+data class LineOptions(val color: RenderColor, val width: Float)
 
 @Serializable
 sealed class RenderOptions
